Logged connection error message before shutting down the executor.
Previously, if an executor failed to connect to the agent, it would
silently shutdown itself without writing an error message to the log.
After we added the support for the domain sockets, a set of potential
failures during `connect` increased. In this patch, we logged
the connection failures to help in debugging.

Review: https://reviews.apache.org/r/72475