Added cmake build option to install build-time module dependencies.
This patch introduces `ENABLE_INSTALL_MODULE_DEPENDENCIES` option
that adds cmake install targets for libraries (including bundled
3rd-party ones) and headers needed to build Mesos modules, similarly
to the option `--enable-install-module-dependencies` in the automake
build.
Note that the layout of installed files is different from the one
produced by automake.

Install targets for stout and libprocess headers are added separately
in the next commits.

This closes #363