Fixed a compilation issue on Windows with os::spawn.
Per MESOS-10137, the call to os::spawn touched here can be seen
as ambiguous by the compiler, given that windows adds a default
argument to os::spawn. Probably we need to remove the default
argument and explicitly implement os::spawn which calls into the
additional windows os::spawn with the extra environment argument.