Fixed a bug in CSI server initialization.
Previously, the CSI server would initialize the service
managers before the auth token was generated, meaning
that requests made by the service managers to an agent
which requires HTTP authentication would fail.

This patch changes the order of initialization so that
the service managers will be initialized with a valid
auth token when necessary.

Review: https://reviews.apache.org/r/72799/