Made UpdateFramework tests not rely on redundant FRAMEWORK_UPDATED.
This patch adjusts UpdateFramework tests so that they do not rely on
an API subscriber receiving the FRAMEWORK_UPDATED event when the
UPDATE_FRAMEWORK call does not change the `FrameworkInfo`.

This is a prerequisite for removing redundant updates and broadcasts
when UPDATE_FRAMEWORK leaves `FrameworkInfo` the same (which typically
happens when the call is used to change suppressed roles set or/and
offer constraints, neither of which is a part of `FrameworkInfo`).

Review: https://reviews.apache.org/r/72823