Consolidated creation and validation of `allocator::Framework` options.
This merges three near-identical pieces of scattered code in SUBSCRIBE
and UPDATE_FRAMEWORK execution paths in the Master that validate
and construct parts of `allocator::FrameworkOptions` (the set of
suppressed roles and the offer constraints filter) into a single
function.

This is a prerequisite to adding validation of offer constraint roles.

Review: https://reviews.apache.org/r/72955