Backported a boost-mpl commit to ignore GCC's -Wparentheses.
In order to allow us to build with `-Werror`.
Since boost is quite a large dependency and the change is so small, it
felt much simpler and less risky than updating boost just to silence
this warning.

See https://github.com/boostorg/mpl/pull/34 for the upstream fix.

This together with https://github.com/apache/mesos/pull/392 allows to
build with `-Werror` using this somewhat recent gcc:
```
gcc (Debian 10.2.1-6) 10.2.1 20210110
```

@asekretenko @qianzhangxa

This closes #393