Fixed NestedMesosContainerizerTest hangs on errors.
Those tests would use a named pipe to synchronize with the task being
started.
The problem is that if the task fails to start, reading from the pipe
would block indefinitely, making the tests just hang.
We could update the code to use a read with a timeout, however it's a
bit fiddly and it's simpler to just use the presence as a regular file
as a barrier.