Explicitely fail content length fetch in case of error
Fetching artifacts is done in two steps:

- use a HEAD request to get size of the remote object
- allocate needed space on disk (by eventually removing old cached data)
- fetch artifact
- validate size of fetched artifact and adapt cache behavior if there is
  no match (but it's outside the scope of this PR)

The issue takes place if the remote server does not answer properly
to HEAD request (ie a 404 or a 500), the current code will return the size
of the 404 or 500 page as if it was the size of the artifact.

There is no "real" issue now that #414 is merged, but there could still be some
rare issues if the artifact is significantly larger and available cache is too small
(Mesos would write more on disk than configured for its cache).

This patch will make Mesos fail sooner if HEAD request fails, but most of the time
this is not an issue as the second request (GET) will fail too anyway.