Fixed a crash in Storage Local Resource ProviderProcess.
`StorageLocalResourceProviderProcess::connected` can crash on a check
that the current state is `DISCONNECTED` if the current state is
`READY`, which can happen if the periodic reconciliation runs after
disconnection.

It can be reproduced by running
`ContentType/AgentResourceProviderConfigApiTest.Add/0` in a loop,
preferably with some CPU-intensive workload in the background to affect
the timing.

Update the check to allow `READY` as well.