Added ENABLE_JEMALLOC_ALLOCATOR macro.
Summary:
The macro is used in the agent code to determine if jemalloc was enabled
in the build and provide a custom allocator configuration in the form of
a global malloc_conf variable.

This commit follows from https://github.com/apache/mesos/commit/2836057c8881a62139b5db1f175d3d027bf2e645.
View the commit for more details.

Test Plan:
Modified the agent code to print whether the macro is defined or not
and tested on an integration test host.