Enforce libnl3 >= 3.5 to prevent overflow errors in the port mapping isolator.
libnl 3.5 changes the interface for link rates (bytes/s) from using a 32 bit integer to
a 64 bit integer. Given that modern links can be on the order of 100 GB/s, using
a 32 bit integer introduces a potential for overflow. This change prevents this kind
of overflow from happening.

Link rates are queried through libnl in the port mapping isolator.

Relevant libnl commit:
- https://github.com/thom311/libnl/commit/4cf69a1b7b2518d18e1a62c79a461bca109132fc