Update boost version to 1.81.0 for CentOS 9 support.
Boost 1.65.0 makes use of functions that are deprecated.
Specifically, `std::unary_function` which was deprecated in
C++11 and removed in C++17. As a result, boost cannot be built
on CentOS 9, which by default only supports C++17 (GCC 4.8.5).

Therefore, we're updating boost to a newer version which can be built
using CentOS 9's GCC 11.4.1 compiler.

Tested on CentOS 7 and CentOS 9.