[cgroups2] Introduces an interface to read and write the CPU bandwidth limit.
In cgroups v2, the CPU bandwidth and bandwidth period (duration over
which the bandwidth can be spent) are set in `cpu.max`. This patch
introduces an interface to read and update these values.

A `BandwidthLimit` object is introduced, which represents a snapshot
of the `cpu.max` control file.

---

Note: This stands in contrast to cgroups v1 where the period and
bandwidth were set in separate control files, `cpu.cfs_period_us`
and `cpu.cfs_quota_us` respectively.