[cgroups2] Introduces cgroups2::exists to check if a cgroup exists.
If you call `cgroups2::destroy` on a cgroup that does not exist, it
will throw an error. This is expected. However, it means that the
caller needs to know whether a cgroup exists. This PR introduces
`cgroups2::exists`, allowing the caller find out if a cgroup exists.

This is useful, for example, for test fixtures where cgroups
from a previous run of a test may or may not be cleaned up, depending
on whether or not the test was successful.