[cgroups2] Introduces API to move a process into a cgroup, by PID.
A process can be moved into a cgroup by writing its process id into
the cgroup's `cgroup.procs` file. When a process is moved, all its
threads move with it.

In this PR we introduce `cgroups2::move_process` to add a process to
a cgroup, `cgroups2::cgroup` to check which cgroup a process is a
member of, and test the introduced functionality.

This closes #523