[cgroups2] Introduces the cpu namespace and cpu::weight() API
Introduces the `cpu` namespace for cgroups2, and expose an API to set
the CPU weight for a cgroup.

In cgroups v2, CPU cycles are allocated based on weights that are assigned
to cgroups, using the "cpu.weight" control. Specifically, a cgroup has access
to 100*(weight/total_weight)% of CPU access.

This closes #522