[cgroups2] Hotfix for controllers::enable/available.
Currently when we read the "cgroup.controllers" and
"cgroup.subtree_control" control files we don't trim
the trailing newline and we don't check if the files
are empty.

This resulted in the empty string, "", being treated
as a controller, which was causing a downstream error
when writing to "cgroup.subtree_control".