[cgroups2] Added flag to enable cgroups v2.
Cgroups v2 is disabled by default. If `--enable-cgroups-v2` is set,
then the cgroups v2 requirements and tests will be added to the build.
This includes `src/linux/cgroups2.*` and `src/linux/ebpf.*` and their
tests.

We introduce this because older Linux kernel versions do not support
cgroups2 but we still want to be able to compile Mesos on them.