[cgroups2] Introduces a `Controller` abstraction for cgroups v2 controllers.
NOTE: In cgroups v1 we call "controllers" "subsystems". In cgroups v2,
      we exclusively use the term "controller", which is what is used
      in the Linux documentation for cgroups v2.

For cgroups v1, the `Subsystem` abstraction is used represent a cgroup
controller. `Subsystem`s exist for each of the controllers provided by
the cgroups v1 API.

We do similar for cgroups v2, now introducing the `Controller` abstraction.
The difference between a `Controller` and a `Subsystem` - besides the
name - is that a `Controller` does not have an associated hierarchy. This is
because in cgroups v2, controllers (a.k.a. subsystems) do not need to be
individually mounted. All controllers are "mounted" under the unified
`cgroup2` filesystem that we require be mounted at `/sys/fs/cgroup`.

The `Cgroups2IsolatorProcess` delegates resource isolation requests to
`Controller`s instead of `Subsystem`s.