[cgroups2] Introduced API to listen for OOM events.
Introduces `cgroups2::memory::events::oom` which returns a future
that resolves when the cgroup reaches its memory limit and is
about to fail with an OOM.

In cgroups v1, there was a bespoke notification API. The same
does not exist in cgroups v2, meaning a new mechanism is used
to detect OOM events.

Cgroups v2 provides the control 'memory.events' which contains
key-value pairs of events and the number of times they took
place [1]. For OOMs, we look at the value of the `oom` field.
In `cgroups2::memory::events::oom` we poll, every 100ms, for
changes to 'memory.events' and resolve a future when
`events.oom > 0`, indicating that _"the cgroup’s memory usage
was reached the limit and allocation [is] about to fail."_

This will be used inside of the `MemoryControllerProcess`, an
isolator what will be provided by the `Cgroups2IsolatorProcess`.

[1] https://docs.kernel.org/admin-guide/cgroup-v2.html#memory