[cgroups2] Introduced API to listen for OOM events.
Introduces `cgroups2::memory::events::oom` which returns a future
that resolves when the cgroup reaches its memory limit and allocation
was about to fail.

In cgroups v1, there was a bespoke notification API.

Cgroups v2 provides the 'memory.events' control which contains key-value
pairs of events and the number of times they took place [1]. For OOMs, we
look at the value of the `oom` field. In `cgroups2::memory::events::oom`
we watch for changes to 'memory.events' (via polling every 100ms for now,
and later via inotify) and resolve a future when `events.oom > 0`.

[1] https://docs.kernel.org/admin-guide/cgroup-v2.html#memory

This closes #563