[cgroups2] Error if `--cgroups_limit_swap` is used when cgroups v2 is used.
Mesos does not support limiting swap memory when using cgroups v2.
This is because the cgroups v2 API allows separate control of swap usage
and careful consideration is needed to figure out how to limit swap
usage.

Therefore, when the `--cgroups_limit_swap` flag is provided and
cgroups v2 is used we error during flag validation.

This closes #565