[cgroups2] Don't enable controllers in the leaf cgroup.
We cannot enable all the controllers in the leaf cgroup because
we also put the container process in the leaf cgroup, which violates
the no internal process constraint.

For instance, if we enable the memory controller in the leaf cgroup
and then try and move a process into the leaf cgroup the operation
will fail with EBUSY.

If a container wants to manage their own cgroups, they will need to
move their process into a new cgroup _before_ they enable controllers.