[cgroups2] Add cgroups v2 initialization and teardown logic to tests fixtures.
StartSlave() and similar test-setup functions mounted cgroups v1 hierarchies
and initialized controllers. On cgroups v2 machines, this setup would fail or
result in irregular cgroup setups. As a step towards end-to-end testing for
the MesosContainerizer, we update the Agent test fixtures such that they
work correctly on both cgroups v1 and v2 hosts.