[cgroups2] Supplement missing 'kernel' field in 'memory.stat' with other kernel fields.
The 'kernel' key was introduces to 'memory.stat' in Kernel 5.18 and
therefore may be missing. If it is missing, we set 'kernel' to be the
sum of the other kernel usage fields provided in 'memory.stat'.

As part of this, we add the 'slab' key, a kernel memory usage field,
to the `memory::Stats` structure.

Kernel patch introducing 'kernel':
https://github.com/torvalds/linux/commit/a8c49af3be5f0b4e105ef678bcf14ef102c270be