[cgroups2] Handle missing 'kernel' field in 'memory.stat' on linux < 5.18.
The 'kernel' key was introduced to 'memory.stat' in Kernel 5.18 and therefore
isn't present on older kernels. If it is missing, we set `kernel` to be the
sum of the other kernel usage fields provided in 'memory.stat'. This is an
under-accounting since it doesn't include:

 - various kvm allocations (e.g. allocated pages to create vcpus)
 - io_uring
 - tmp_page in pipes during pipe_write()
 - bpf ringbuffers
 - unix sockets

But it's the best measurement we can provide prior to the 'kernel' stat
being added in 5.18 that catches all of these.

As part of this, we add the 'slab' key (one of the kernel memory usage
fields) to the `memory::Stats` structure.

See kernel patch introducing 'kernel':

https://github.com/torvalds/linux/commit/a8c49af3be5f0b4e105ef678bcf14ef102c270be

This closes #576