[cgroups2] adjust CPU weight values from v1 to v2 default
Modifies the cgroups CPU weights to reflect change
from cpu.shares to cpu.weight.

In v1, cgroups used cpu.shares which has a default of 1024.
In v2, cgroups use cpu.weight which has a default of 100

The range for the cpu.weight is [1,10000], the minimum
weight has been updated to reflect this.

The revocable CPU weight has been scaled down from 10 to 1
to reflect a similar scale to the default.

Review: https://reviews.apache.org/r/74992/