[cgroups2] crash when root folder is not detected when creating cgroups
Based on this ticket (https://issues.apache.org/jira/browse/MESOS-9305)
and the ROOT_CGROUPS_CreateRecursively test in CgroupsIsolatorTest,
there seems to be a possibility that the root folder may be deleted and
new cgroups cannot be properly created.

In v1, this was addressed by enabling recursively creating the groups.

In v2, since we make use of cgroup.subtree_control to determine a cgroup
and its descendents' access to controllers, we cannot recover this effectively
if the root folder is deleted, so we cant just recursively create the folders.

Hence, we elected to crash if the root folder is not found, as it will allow
us to restart and go through the logic that takes care of setting all the
values inside cgroup.subtree_control again.

Review: https://reviews.apache.org/r/74995/