[cgroups2] Introduce the PerfEventControllerProcess.
Introduces the controller process for perf event which was also present
in cgroups1. The controller is automatically enabled, and should not be
visible inside the cgroups.controllers file in the root cgroup.

As a consequence, we will not be able to manually enable or disable this
controller via writing to the cgroup.subtree_control file.

References:

* perf_event section in https://docs.kernel.org/admin-guide/cgroup-v2.html
* slide 34 in https://man7.org/conf/ndctechtown2021/cgroups-v2-part-1-intro-NDC-TechTown-2021-Kerrisk.pdf

Review: https://reviews.apache.org/r/74997/