[agent] Add executor_id / framework_id query parameters in /containers.
We now allow filtering by framework ID and executor ID in addition to
the original functionality of filtering by container ID.

Please note that the /containers endooint only allows select combinations
of these query parameter fields to be populated at once. We will return a
failure if we see that the combination of query paramters is invalid.

We currently accept:
* no query parameters
* only container id
* only framework id
* only framework id and executor id

Review: https://reviews.apache.org/r/75009/