[cgroups2] Remove ENABLE_CGROUPS_V2 ifdefs.
This commit removes the sections where ENABLE_CGROUPS_V2 is used
to determine the compiled code. Any need to determine whether or not
cgroups2 is used will be satisfied using the cgroups2::mounted() function.

This guard was only in place temporarily to avoid breaking our CI
while we figured out how to ensure that all of the CI docker images
have the header.

Review: https://reviews.apache.org/r/75021/