[mesos-build] Address dependency issues in centos-7 / ubuntu-20.04.
In CentOS 7, the linux/amd64 base image is missing ebpf fields such as
BPF_PROG_TYPE_CGROUP_DEVICE, which prevents jenkins from building mesos
now that the ENABLE_CGROUPS_V2 macro has been removed as
/usr/include/linux/bpf.h is missing the fields required by ebpf.h.

We installed dependencies from kernel-ml so that we can have newer
headerfiles in /usr/include/linux, which should help us compile the
mesos code in jenkins.

For ubuntu, there is a dependency on installing jdk-11 instead of
the old jdk-8 which is preventing some builds which pull the ubuntu
image to build.

As part of this change, both the CentOS 7 and ubuntu:20.04 will
need to be rebuilt and uploaded to dockerhub for jenkins.

Review: https://reviews.apache.org/r/75035/