[mesos-build] Remove setting of environment variable in dockerfiles.
Setting an environment variable as PYTHON_VERSION caused an unexpected
behavior in jenkins as the configure.ac script checked for that exact
environment variable.

PYTHON_VERSION has been renamed and set as an ARG in the dockerfile
so that it will not persist after the build.

Review: https://reviews.apache.org/r/75036/