[routing] Change link::setMAC to return Try<Nothing>.
We have noticed that our code does not treat the setMAC bool return value
differently based on whether it returns true or false. As such, we are
changing the return type to return Nothing so that we either return Error
or Nothing, rather than Error or True or False.

As a consequence of this we are also removing the special case of returning
False but not Error when we get ENODEV from ioctl.

Review: https://reviews.apache.org/r/75056/