[port mapping isolator] Work around apparent MAC address kernel bug.
It seems that there are scenarios where, when using the port mapping isolator,
mesos containers sometimes cannot communicate with the mesos agent as the MAC
address of the veth interface is set incorrectly, leading to dropped packets
by the kernel. This was discovered with the use of tcpdump (which reveals that
the kernel marks the packets as destined for another host), and the latter of
which reveals that the kernel is indeed dropping the packets due to this. We
then found that when we set the mac address on the veth interface, it sometimes
does not "stick" despite ioctl returning successfully.

Observed scenarios with incorrectly assigned MAC addresses:

1. After setting the mac address: ioctl returns the correct MAC address, but
   net::mac returns an incorrect MAC address (different from the original!)
2. After setting the mac address: both ioctl and net::mac return the same MAC
   address, but are both wrong (and different from the original one!)
3. After setting the mac address: there are no cases where ioctl or net::mac
   come back with the same MAC address as before we set the address.
4. Before we set the mac address: there is a possibility that ioctl and
   net::mac results disagree with each other!
5. There is a possibility that the MAC address we set ends up overwritten by
   a garbage value after setMAC has already completed and checked that the
   mac address was set correctly. Since this error happens after this
   function has finished, we cannot log nor detect it in setMAC because we
   have not yet studied at what point this occurs.

Notes:

1. We have observed this behavior only on CentOS 9 systems at the moment,
   CentOS 7 systems under various kernels do not seem to have the issue
   (which is quite strange if this was purely a kernel bug).
2. We have tried kernels 5.15.147, 5.15.160, 5.15.161, all of these have
   this issue on CentOS 9.

This patch adds a workaround for this bug, which is to check that the MAC
address is set correctly after the ioctl call, and retry the address setting
if necessary. In our testing, this workaround appears to workaround scenarios
(1), (2), (3), and (4) above, but it does not address scenario (5).

See MESOS-10243 for additional details, follow-ups.

Review: https://reviews.apache.org/r/75057/