[build] Fix make distcheck for ubuntu 20.04.
Fixes the makefiles for ubuntu 20.04 so that make distcheck works properly with
its protobuf dependencies generated as part of make distcheck.

The reason that a change was needed in the makefile is because the upgrade from
ubuntu 16.04 to 20.04 also caused the automake version to be updated when
dependencies were being installed during docker build.

The change in the automake version created slight changes in the generated
makefile. Speficically, the distcheck on the new automake-generated makefile now
depends on `BUILT_SOURCES` which causes an error as the CSI protobuf files are
not ready when distcheck is called. So we add the csi build stamps to
`BUILT_SOURCES` to ensure that the protobuf files will be ready when distcheck's
dependencies are made.

The additional chmod change for java is because for some reason when distcheck
attempts to build the mesos-1.12.0.jar from its created distribution, some
folders are missing write permissions, causing the build to terminate.

Review: https://reviews.apache.org/r/75062/