[build] Fix libevent-enabled cmake builds on ubuntu 20.04.
Our current libevent-enabled cmake builds cannot complete on jenkins as it
gets the 'incomplete definition of type 'struct bio_st'' error.

This is because the upgrade to ubuntu 20.04 also upgraded our openSSL
version from 1.0.2 to 1.1.1, which breaks compatibility with libevent 2.1.5
that was previously used.

A compatibility patch for openSSL 1.1+ was released with libevent 2.1.7, but
the closest tarball that includes a CMakeLists.txt file is 2.1.9, which is
what we will upgrade to.

With the new libevent library, builds are able to complete using cmake and
with libevents enabled. But it still sees the test failures we see on other
builds (such as autotools) and operating system (CentOS 7).

We also need to link libevent_pthread and libevent_openssl with libprocess,
otherwise we will get errors like:

```
ld: 3rdparty/libprocess/src/libprocess.so: undefined reference to bufferevent_openssl_get_ssl
```

Review: https://reviews.apache.org/r/75070/