[build] Use clang-14 for non ubuntu 16.04 targets in docker-build.sh.
As we migrate to ubuntu 22.04, clang-10 is no longer available via
apt-get install. As such, we will move to clang-14, which should allow
us to run the docker-build.sh file with OS equal to ubuntu:22.04.
This should allow coverity bot to compile and return to normal.

Review: https://reviews.apache.org/r/75091/