[cgroups2] Introduce a device manager.
This change introduces the DeviceManager to help facilitate device
access management in cgroups2 via ebpf program file changes. This
centralization is needed since we no longer have control files to
leverage as persistence for agent recovery, so we a component that
keeps track of allow/deny device access information and re-configures
the ebpf program for the cgroup.

Device requests can be made to the manager by calling `configure` or
`reconfigure`. Note that `configure` should only be used when setting
up a cgroup's device access, i.e. it has not requested any device to
be allowed/denied before.

In addition, `reconfigure` cannot be used to add deny entries
containing wildcards.

This manager will be made available to all controllers under the
cgroups2 isolator, and the GPU isolator.

Review: https://reviews.apache.org/r/75006/