[cgroups] Add helper to find overlapping device access.
Currently, we have to directly compare member variables to see if one
Access object would overlap that of another, which isn't very clear
to people that would be reading the code.

We add a helper to abstract away the logic to see if the accesses
specified in one Access instance would overlap with that of another.

Review: https://reviews.apache.org/r/75107/